@@ -5,8 +5,8 @@
  */
 
 class DemoList extends Object {
-  List<DemoItem> list;
-  int total;
+  List<DemoItem?>? list;
+  int? total;
 
   DemoList({
     this.list,
@@ -28,9 +28,9 @@ class DemoList extends Object {
 }
 
 class DemoItem extends Object {
-  String id;
-  String name = '';
-  String type = '';
+  String? id;
+  String? name = '';
+  String? type = '';
 
   DemoItem({this.id, this.name, this.type});
 
