@@ -8,7 +8,7 @@ library fair;
 
 export 'package:fair_version/fair_version.dart';
 
-// export 'src/annotation.dart';
+export 'src/annotation.dart';
 export 'src/app.dart';
 export 'src/dialog.dart';
 export 'src/experiment/sugar.dart';
